Safari-609.1.11.1/LayoutTests/http/tests/contentextensions/make-https-expected.txt

CONSOLE MESSAGE: line 29: Content blocker promoted URL from http://127.0.0.1/promote to https
CONSOLE MESSAGE: line 29: Content blocker promoted URL from http://127.0.0.1/promote to https
CONSOLE MESSAGE: line 29: Content blocker promoted URL from http://127.0.0.1/promote to https
CONSOLE MESSAGE: line 29: Content blocker promoted URL from http://promote/ to https
Blocked access to external URL https://promote/
URLs with http, port 80, and promote in the url should be promoted to https with port 443. Four URLs should be promoted as should be shown by four CONSOLE MESSAGE lines.
